+++ CHANGELOG.md
@@ -1,3 +1,121 @@
+# [4.4.0](https://github.com/aeternity/aepp-sdk-js/compare/4.3.0...4.4.0) (2019-08-09)
+
+
+### Bug Fixes
+
+* **Package:** update commander to version 3.0.0 ([#597](https://github.com/aeternity/aepp-sdk-js/issues/597)) ([9aaa05c](https://github.com/aeternity/aepp-sdk-js/commit/9aaa05c))
+* **Contract:** Fix dry-run without account ([227fc5c](https://github.com/aeternity/aepp-sdk-js/pull/599/commits/227fc5c9d8369e1c3a3fbba2cf62f0495bcf08ae))
+
+### Features
+
+* **Contract:** add ability to use call-static/dry-run without keyPair ([#577](https://github.com/aeternity/aepp-sdk-js/issues/577)) ([c38edd9](https://github.com/aeternity/aepp-sdk-js/commit/c38edd9))
+* **AE:** Add ability to make operation on specific account using `onAccount` option.
+   ``` 
+    Exm: await client.spend(1, receiver, { onAccount: 'PUBLIC_KEY' })
+    Add `onAccount` to `AENS`, `Contract`, `Oracle`.
+    Add tests for using specific account to Contract, ACI, Account.
+  ```
+* **JSON:**: Add serialization to JSON for bigNumbers
+* **MemoryAccount:** Add validation of `keypair`  ([#594](https://github.com/aeternity/aepp-sdk-js/issues/594)) ([b8c2b20](https://github.com/aeternity/aepp-sdk-js/commit/b8c2b20))
+* **State Channels:** persist connection by pinging every 10 seconds ([#571](https://github.com/aeternity/aepp-sdk-js/issues/571)) ([a70f919](https://github.com/aeternity/aepp-sdk-js/commit/a70f919))
+
+
+
+# [4.3.0](https://github.com/aeternity/aepp-sdk-js/compare/2.4.0...4.3.0) (2019-08-05)
+
+
+### Bug Fixes
+
+* **State Channels:** Fix onChainTx event params ([#566](https://github.com/aeternity/aepp-sdk-js/issues/566)) ([11c85eb](https://github.com/aeternity/aepp-sdk-js/commit/11c85eb))
+* **State Channels:** Fix websocket url ([#558](https://github.com/aeternity/aepp-sdk-js/issues/558)) ([33c1fd8](https://github.com/aeternity/aepp-sdk-js/commit/33c1fd8))
+* **Swagger:** Pass query params in case of get request ([#570](https://github.com/aeternity/aepp-sdk-js/pull/570))
+
+### Code Refactoring
+
+* **State Channel:** Do not include white space for outgoing websocket messages ([#559](https://github.com/aeternity/aepp-sdk-js/pull/559))
+
+
+### Features
+
+* **ACI:** Implement sophia `variant` type ([#567](https://github.com/aeternity/aepp-sdk-js/issues/567)) ([8505dcf](https://github.com/aeternity/aepp-sdk-js/commit/8505dcf))
+* **Contract:** add ability to use call-static/dry-run without keyPair ([#577](https://github.com/aeternity/aepp-sdk-js/issues/577)) ([c38edd9](https://github.com/aeternity/aepp-sdk-js/commit/c38edd9))
+* **NodePool:** Implement NodePool stamp ([#574](https://github.com/aeternity/aepp-sdk-js/issues/574)) ([674166c](https://github.com/aeternity/aepp-sdk-js/commit/674166c))
+* **State Channel:** make state channel compatible with aeternity@4.0.0 ([#568](https://github.com/aeternity/aepp-sdk-js/issues/568)) ([0d0e09b](https://github.com/aeternity/aepp-sdk-js/commit/0d0e09b))
+* **TxBuilder:** Add helper for producing tx hash ([#579](https://github.com/aeternity/aepp-sdk-js/issues/579)) ([e1b405e](https://github.com/aeternity/aepp-sdk-js/commit/e1b405e))
+* **AE** Make all AE stamps composed with Accounts ([#581](https://github.com/aeternity/aepp-sdk-js/pull/581))
+
+
+### Docs
+
+* **ACI** Add some additional clarification to `getContractInstance`
+
+
+
+# [4.2.0](https://github.com/aeternity/aepp-sdk-js/compare/2.4.0...4.2.0) (2019-07-15)
+
+
+### Bug Fixes
+
+* **package:** update libsodium-wrappers-sumo to version 0.7.5 ([#541](https://github.com/aeternity/aepp-sdk-js/issues/541)) ([956ed75](https://github.com/aeternity/aepp-sdk-js/commit/956ed75))
+* **rpc-server:** Fix type 'object' check ([#526](https://github.com/aeternity/aepp-sdk-js/issues/526)) ([48c42e4](https://github.com/aeternity/aepp-sdk-js/commit/48c42e4))
+
+
+### Code Refactoring
+
+* **swagger:** Speedup initialisation
+* **AENS:** Remove unused param from claim method
+* **AENS:** Fix exception if not waiting for mining(claim)
+* **Test:** Add test for contract namespaces
+
+### Features
+
+* **Node:** Add 4.0.0 node compatibility
+* **Compiler:** Add compatibility with compiler 3.2.0
+* **Channel:** Implement GA awareness of State Channels
+
+
+
+# [4.1.0](https://github.com/aeternity/aepp-sdk-js/compare/4.0.1...4.1.0) (2019-06-22)
+
+
+### Bug Fixes
+
+* **Node:** Do not throw error if `internalUrl` not provided. Instead use `url` ([#503](https://github.com/aeternity/aepp-sdk-js/issues/503)) ([053faae](https://github.com/aeternity/aepp-sdk-js/commit/053faae))
+* **TXBuilder:** Fix payload serialization if you try to unpack and pack tx. ([#498](https://github.com/aeternity/aepp-sdk-js/issues/498)) ([73552e5](https://github.com/aeternity/aepp-sdk-js/commit/73552e5))
+* **TxValidator:** Fix validation of state channel open transaction ([#496](https://github.com/aeternity/aepp-sdk-js/issues/496)) ([325cc90](https://github.com/aeternity/aepp-sdk-js/commit/325cc90))
+
+
+### Features
+
+* **ACI:** Refactor ACI module. Split to separated files. ([#505](https://github.com/aeternity/aepp-sdk-js/issues/505)) ([fb7bc00](https://github.com/aeternity/aepp-sdk-js/commit/fb7bc00))
+* **Selector:** If default account `address` not provided use the first
+* **ACI:** Handle ACI without init function
+* **ACI:** Automatically decide to send transaction on-chai or call-static.
+Add `options` object like last arguments of generate fn under `instance.methods`
+  ```
+  const instance = await client.getContractInstance(source)
+  // Deploy contract
+  await.contract.init(100, 'test', options)
+  //or
+  await.contract.deploy([100, 'test], options)
+  // Call function
+  const result = await instance.call('sum', [2, 5], options)
+  //
+    // Automatically decide to send tx on-chain or call-static(dry-run) base on if function stateful or not
+    const result = await instance.methods.sum(2, 5, options)
+    // Manually make on-chain
+    const result = await instance.methods.sum.send(2, 5, options)
+    // Manually make call-static
+    const result = await instance.methods.sum.get(2, 5, options)
+  //
+  ```
+
+### Docs
+
+* **Usage:**: Add instructions about how to include directly the `SDK` in a `html page`
+
+
+
 ## [4.0.1](https://github.com/aeternity/aepp-sdk-js/compare/2.4.0...4.0.1) (2019-06-13)